+++ src/app/features/settings/settings.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { RouterOutlet, RouterLink,  RouterModule } from '@angular/router';
 import { MatListModule } from '@angular/material/list'; 
-import {NavbarComponent} from '../../shared/components/navbar/navbar.component'
 import { MatDividerModule } from '@angular/material/divider';
 import { MatIcon } from '@angular/material/icon';
 import { TranslatePipe } from '@ngx-translate/core';
@@ -10,7 +9,7 @@ import { TranslatePipe } from '@ngx-translate/core';
 @Component({
   selector: 'app-settings',
   standalone: true,
-  imports: [NavbarComponent,TranslatePipe, MatIcon,MatSidenavModule, RouterModule, MatDividerModule, RouterOutlet, RouterLink, MatListModule],
+  imports: [TranslatePipe, MatIcon,MatSidenavModule, RouterModule, MatDividerModule, RouterOutlet, RouterLink, MatListModule],
   templateUrl: './settings.component.html',
   styleUrl: './settings.component.scss'
 })