+++ src/environments/environment.ts
@@ -9,7 +9,10 @@ export const environment = {
   base_url: 'http://localhost:8081',
   wallet_url: 'http://localhost:4200',
   wallet_url_test: 'http://localhost:4200',
-  knowledgebase_url: "https://knowledgebase.dome-marketplace-sbx.org/",
+  knowledge:{
+    base_url: "https://knowledgebase.dome-marketplace-sbx.org/",
+    wallet_path: "books/dome-digital-wallet-user-guide"
+  } ,
   profile: "lcl", //values: 'lcl', 'sbx', 'test', 'production'
   procedures: '/api/v1/procedures',
   save_credential: '/vci/v1/issuances',
@@ -22,7 +25,7 @@ export const environment = {
       secondary:'#2cb6b2',
       secondary_contrast:'#dde6f6'
     },
-    logo_src:"assets/logos/no-image.png",
+    logo_src:"assets/logos/logo-dome-issuer-reduced.png",
     favicon_src:"assets/icon/favicon.png"  
   }
   //firma_credential: '/api/v1/sign-credential' The`firma_credential` variable has been commented out as it was initially intended for the signature functionality,which remains incomplete. This configuration is currently unnecessary for the existing flows but is expected to be reintroduced in the future when the related use case is implemented.